refactor: app, entities folders
