+++ src/app/providers/index.tsx
@@ -1,0 +1,1 @@
+export { ReduxProvider } from './ReduxProvider'
