+++ src/widgets/EmailSentPopUp/ui/EmailSentPopUpWidget.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react'
 
-import { selectAuthUser } from '@/entities/auth/authSlice'
+import { selectAuthUser } from '@/entities/auth/model/authSlice'
 import { CloseIcon } from '@/shared/assets'
 import { useAppSelector, useTranslation } from '@/shared/model'
