+++ src/widgets/signUp/ui/SignUpWidget.tsx
@@ -10,8 +10,8 @@ import { SignUpAuth } from '../signUpAuth/SignUpAuth'
 import styles from './SignUpWidget.module.scss'
 
 import { AppDispatch } from '@/app/appStore'
-import { useRegistrationMutation } from '@/entities/auth/authApi'
-import { setUser } from '@/entities/auth/authSlice'
+import { useRegistrationMutation } from '@/entities/auth'
+import { setUser } from '@/entities/auth/model/authSlice'
 import { AUTH_URLS } from '@/shared'
 import { GithubIcon, GoogleIcon } from '@/shared/assets'
 import { consoleErrors, useTranslation } from '@/shared/model'
